requirements.txt
streamlit==0.88.0
biopython==1.79